+++ sdk/src/main/java/com/herewhite/sdk/ResultCaller.java
@@ -1,0 +1,5 @@
+package com.herewhite.sdk;
+
+public interface ResultCaller<T> {
+    void call(T result);
+}