+++ sdk/src/main/java/com/herewhite/sdk/WhiteSdkConfiguration.java
@@ -28,6 +28,7 @@ public class WhiteSdkConfiguration extends WhiteObject {
     private boolean log = false;
     private RenderEngineType renderEngine = RenderEngineType.canvas;
     private boolean enableInterrupterAPI = false;
+    private boolean enableSlideInterrupterAPI = false;
     private boolean preloadDynamicPPT = false;
     private boolean routeBackup = false;
     private boolean userCursor = false;
@@ -45,6 +46,10 @@ public class WhiteSdkConfiguration extends WhiteObject {
      * pptParams 动态 ppt 专用参数
      */
     private PptParams pptParams = new PptParams();
+    /**
+     * SlideApp 的配置项
+     */
+    private SlideAppOptions slideAppOptions = new SlideAppOptions();
     private HashMap<String, String> fonts;
     private boolean enableImgErrorCallback;
     /**
@@ -500,6 +505,32 @@ public class WhiteSdkConfiguration extends WhiteObject {
         this.useMultiViews = useMultiViews;
     }
 
+    public SlideAppOptions getSlideAppOptions() {
+        return slideAppOptions;
+    }
+
+    public void setSlideAppOptions(SlideAppOptions slideAppOptions) {
+        this.slideAppOptions = slideAppOptions;
+    }
+
+    public boolean isEnableSlideInterrupterAPI() {
+        return enableSlideInterrupterAPI;
+    }
+
+    /**
+     * 开启/关闭 SlideApp 拦截替换功能。
+     * <p>
+     * 该方法可以开启或关闭 SlideApp 资源的拦截功能。
+     * 如果开启，在加载 url 资源时，SlideApp 会拦截图片并触发 {@link com.herewhite.sdk.window.SlideListener#slideUrlInterrupter(String, ResultCaller)} 回调，你可以在该回调中替换图片的地址。
+     *
+     * @param enableSlideInterrupterAPI 是否开启 SlideApp 资源拦截和替换功能：
+     *                             - `true`：开启。
+     *                             - `false`：（默认）关闭。
+     */
+    public void setEnableSlideInterrupterAPI(boolean enableSlideInterrupterAPI) {
+        this.enableSlideInterrupterAPI = enableSlideInterrupterAPI;
+    }
+
     /**
      * 白板上绘画的渲染模式。
      *
@@ -594,5 +625,29 @@ public class WhiteSdkConfiguration extends WhiteObject {
             this.useServerWrap = useServerWrap;
         }
         /// @endcond
+    }
+
+    // 互动白板 SlideApp 的配置项。
+    public static class SlideAppOptions extends WhiteObject {
+        // 是否显示渲染错误
+        private boolean showRenderError = false;
+        // 是否开启调试模式
+        private boolean debug = false;
+
+        public boolean isShowRenderError() {
+            return showRenderError;
+        }
+
+        public void setShowRenderError(boolean showRenderError) {
+            this.showRenderError = showRenderError;
+        }
+
+        public boolean isDebug() {
+            return debug;
+        }
+
+        public void setDebug(boolean debug) {
+            this.debug = debug;
+        }
     }
 }